pyproject.toml
# SPDX-FileCopyrightText: Florian Maurer
#
# SPDX-License-Identifier: AGPL-3.0-or-later

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "emarketlib"
version = "0.0.1"
description = "Energy market toolbox"
authors = [{ name = "Florian Maurer", email = "maurer@fh-aachen.de"}]
license = {text = "AGPL-3.0-or-later"}
readme = "README.md"
keywords = ["agent based simulation", "energy market", "market simulation", "simulation"]

classifiers=[
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Framework :: AsyncIO",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.10"
dependencies = [
    "argcomplete >=3.1.4",
    "mango-agents >=1.2",
    "numpy >=1.26.4",
    "python-dateutil >=2.8.2",
    "pandas >=2.0.0",
    "pyyaml >=6.0.2",
]

[project.optional-dependencies]
test = [
    "ruff >=0.4.9",
    "mypy >=1.1.1",
    "matplotlib >=3.7.2",
    "pytest >=7.2.2",
    "pytest-cov >=4.1.0",
    "pytest-asyncio >=0.21.1",
]

[project.urls]
homepage = "https://github.com/maurerle/emarketlib"
repository = "https://github.com/maurerle/emarketlib"


[tool.setuptools]
packages = ["emarketlib"]

[tool.ruff]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "G", "PIE"]
ignore = ["E501", "G004", "E731"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "I001", # allow unsorted imports in __init__.py
    "F401", # allow unused imports in __init__.py
]
"tests/*" = [
    "S101", # allow assert statements for pytest
    "ARG",  # allow unused arguments for pytest fixtures
    "F841", # allow unused local variables
]

[tool.pytest]
testpaths = "tests"

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "ignore:Call to deprecated create function FieldDescriptor",
    "ignore:Call to deprecated create function Descriptor",
    "ignore:Call to deprecated create function FileDescriptor",
    "ignore:Call to deprecated create function EnumValueDescriptor",
    "ignore:Call to deprecated create function EnumDescriptor",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
addopts = "--ignore=assume-db"

[tool.coverage.run]
omit = ["tests/*"]
source = ["assume"]

[tool.coverage.report]
# https://stackoverflow.com/a/5850364
exclude_lines = ["if __name__ == .__main__.:"]

[tool.mypy]
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
# output
show_error_context = true
show_column_numbers = true
show_error_codes = true
pretty = true
ignore_missing_imports = true
